server/requirements.txt
aiortc
aiohttp
numpy
onnxruntime
Pillow